+++ tests/Unit/Public/Assert-ObjectMethod.Tests.ps1
@@ -333,7 +333,7 @@ Describe 'Assert-ObjectMethod' {
                                 if ($null -eq $currentObject) {
                                     $message = $script:localizedData.Assert_ObjectMethod_ActualIsNull
                                     if ($Because) {
-                                        $message += " {0} $Because" -f $script:localizedData.Assert_ObjectMethod_Because
+                                        $message += " {0} $Because" -f $script:localizedData.Common_WordBecause
                                     }
                                     throw [Pester.Factory]::CreateShouldErrorRecord($message, 'test', 1, 'test', $true)
                                 }
@@ -372,7 +372,7 @@ Describe 'Assert-ObjectMethod' {
                     if ($null -eq $Actual) {
                         $message = $script:localizedData.Assert_ObjectMethod_ActualIsNull
                         if ($Because) {
-                            $message += " {0} $Because" -f $script:localizedData.Assert_ObjectMethod_Because
+                            $message += " {0} $Because" -f $script:localizedData.Common_WordBecause
                         }
                         throw [Pester.Factory]::CreateShouldErrorRecord($message, 'test', 1, 'test', $true)
                     }